Fix: list my packs + packs
