+++ cd2024bfs2g1-api/src/main/java/com/campusdual/cd2024bfs2g1/api/core/service/pack/IPackService.java
@@ -12,6 +12,9 @@ public interface IPackService {
 
     EntityResult packClientQuery(Map<String, Object> keysValues, List<String> attributes) throws OntimizeJEERuntimeException;
 
+    EntityResult allPacksQuery(Map<String, Object> keysValues, List<String> attributes) throws OntimizeJEERuntimeException;
+
+
     EntityResult packInsert(Map<String, Object> attrMap) throws OntimizeJEERuntimeException;
 
     EntityResult packUpdate(Map<String, Object> attrMap, Map<String, Object> keyMap) throws OntimizeJEERuntimeException;
