--- cd2024bfs2g1-frontend/src/main/ngx/src/app/main/pack/pack-detail/pack-detail.component.css
@@ -1,147 +1,0 @@
-.page {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    width: 100%;
-}
-
-.mainCard {
-    display: flex;
-    justify-content: center;
-    height: auto;
-    width: 1100px;
-}
-
-
-.main-container {
-
-    min-width: 700px;
-    align-content: center;
-    justify-content: center;
-    flex-direction: row;
-    margin-left: 50px;
-    margin-right: 50px;
-    margin-top: 10px;
-
-}
-
-.name-container {
-    margin-top: 20px;
-    margin-bottom:20px;
-}
-
-.data-container {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    width: 100%;
-    
-}
-
-.div-image-container {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    width: 100%;
-    height: 400px;
-    margin-bottom: 30px;
-    align-items: center;
-
-}
-
-.div-active-pack-container{
-    width: 80%;
-    text-align: center;
-    margin: auto;
-}
-
-.row-image-container {
-    display: block;
-    flex-direction: row;
-    justify-content: center;
-    width: 100%;
-    padding: 20px;
-}
-
-.image {
-    width: 100%;
-    height: auto;
-    max-width: 100%; 
-    max-height: 100%;
-    border-radius: 10px;
-    object-fit: cover;
-}
-
-.image-title{
-    position: absolute;
-    justify-content: center;
-    background-color: #84A45A;
-    color: #E4EBF2;
-    border-radius: 10px;
-    padding-left: 10px;
-    padding-right: 10px;
-    
-}
-
-.active-title-pack{
-  border-top-left-radius: 20px;
-  border-bottom-right-radius: 20px;
-  background-color: rgb(187, 9, 9);
-  color:white;
-  height: 35px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 20px;
-
-}
-
-.left, .right {
-    flex: 1;
-    padding: 25px;
-    background-color: transparent;
-        mat-icon{
-            color:#84A45A;
-        
-    }
-    
-}
-
-.desc {
-    margin-left: 20px;
-}
-
-.o-rows {
-    background-color: transparent;
-    
-}
-
-.list-with-spaces {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-around;
-    width: 90%;
-    row-gap: 10px;
-    max-height: 280px;
-    overflow-y: auto;
-}
-
-
-.o-text-input {
-    outline: transparent;
-    width: 100%;
-    height: 100%;
-    margin-top: 15px;
-    
-}
-
-.o-grid-items {
-    width: 100%;
-    height: 100%
-}
-
-
-
-
-
-
