--- cd2024bfs2g1-frontend/src/main/ngx/src/app/main/pack/pack-detail/pack-detail.component.ts
@@ -1,32 +1,0 @@
-import { Component, ViewChild } from '@angular/core';
-import { DomSanitizer } from '@angular/platform-browser';
-import { Router } from '@angular/router';
-import { OFormComponent, OTableComponent } from 'ontimize-web-ngx';
-
-@Component({
-  selector: 'app-pack-detail',
-  templateUrl: './pack-detail.component.html',
-  styleUrls: ['./pack-detail.component.css']
-})
-export class PackDetailComponent {
-  @ViewChild('accountCustomerTable') accountTable: OTableComponent;
-  @ViewChild('form') form: OFormComponent;
-
-  constructor(
-    protected sanitizer: DomSanitizer,
-    private router: Router,
-  ) {
-  }
-
-	ngOnInit(): void {
-  }
-
-  public getImageSrc(base64: any): any {
-    return base64 ? this.sanitizer.bypassSecurityTrustResourceUrl('data:image/*;base64,' + base64) : './assets/images/no-image-transparent.png';
-  }
-
-  public openPacks(): void {
-    this.router.navigate(['main/pack']);
-  }
-
-}
