+++ cd2024bfs2g1-frontend/src/main/ngx/src/app/main/pack/pack-home/pack-home.component.html
@@ -1,4 +1,4 @@
-<o-grid #packGrid attr="packGrid" title="PACKS" service="packs" entity="pack" keys="P.pck_id"
+<o-grid #packGrid attr="packGrid" title="PACKS" service="packs" entity="allPacks" keys="P.pck_id"
     columns="P.pck_id;pck_name;pck_description;pck_date_begin;pck_date_end;pck_price;pck_participants;gui_c_name;pcs_id;img_code"
     query-rows="16" page-size-options="4;8;16;24" orderable="true" cols="3"
     sortable-columns="pck_name" insert-button="true" refresh-button="true"
