+++ cd2024bfs2g1-model/src/main/java/com/campusdual/cd2024bfs2g1/model/core/service/MerchantService.java
@@ -2,14 +2,19 @@ package com.campusdual.cd2024bfs2g1.model.core.service;
 
 import com.campusdual.cd2024bfs2g1.api.core.service.IMerchantService;
 import com.campusdual.cd2024bfs2g1.model.core.dao.MerchantDao;
+import com.campusdual.cd2024bfs2g1.model.core.dao.UserDao;
 import com.campusdual.cd2024bfs2g1.model.core.dao.UserRoleDao;
 import com.ontimize.jee.common.dto.EntityResult;
 import com.ontimize.jee.common.exceptions.OntimizeJEERuntimeException;
+import com.ontimize.jee.common.services.user.UserInformation;
 import com.ontimize.jee.server.dao.DefaultOntimizeDaoHelper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Lazy;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -66,6 +71,29 @@ public class MerchantService implements IMerchantService {
     @Override
     public EntityResult merchantDelete(Map<String, Object> keyMap) throws OntimizeJEERuntimeException {
         return this.daoHelper.delete(this.merchantDao, keyMap);
+    }
+
+    /**
+     * Returns merchant id from logged user
+     *
+     * @return merchant_id
+     */
+    public int getMerchantId() {
+        Object merchant = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        int userId = (int) ((UserInformation) merchant).getOtherData().get(UserDao.USR_ID);
+
+
+        List<String> qKeys = new ArrayList<String>();
+        qKeys.add(MerchantDao.MERCHANT_ID);
+
+        Map<String, Object> emptyMap = new HashMap<>();
+        emptyMap.put("M." + UserDao.USR_ID, userId);
+
+
+        EntityResult merchantEr = merchantQuery(emptyMap, qKeys);
+        ArrayList<Integer> al = (ArrayList<Integer>) merchantEr.get(MerchantDao.MERCHANT_ID);
+        return al.get(0);
     }
 
 }
