+++ cd2024bfs2g1-model/src/main/java/com/campusdual/cd2024bfs2g1/model/core/service/pack/PackService.java
@@ -1,31 +1,25 @@
 package com.campusdual.cd2024bfs2g1.model.core.service.pack;
 
-import com.campusdual.cd2024bfs2g1.api.core.service.IPackBookingService;
 import com.campusdual.cd2024bfs2g1.api.core.service.pack.IPackService;
 import com.campusdual.cd2024bfs2g1.model.core.dao.ClientDao;
 import com.campusdual.cd2024bfs2g1.model.core.dao.ImageDao;
-import com.campusdual.cd2024bfs2g1.model.core.dao.MerchantDao;
-import com.campusdual.cd2024bfs2g1.model.core.dao.UserDao;
 import com.campusdual.cd2024bfs2g1.model.core.dao.business.GuideCitiesDao;
 import com.campusdual.cd2024bfs2g1.model.core.dao.pack.PackDao;
 import com.campusdual.cd2024bfs2g1.model.core.service.ClientService;
 import com.campusdual.cd2024bfs2g1.model.core.service.ImageService;
-import com.campusdual.cd2024bfs2g1.model.core.service.PackBookingService;
 import com.campusdual.cd2024bfs2g1.model.core.service.business.GuideCitiesService;
 import com.ontimize.jee.common.dto.EntityResult;
 import com.ontimize.jee.common.exceptions.OntimizeJEERuntimeException;
-import com.ontimize.jee.common.services.user.UserInformation;
 import com.ontimize.jee.server.dao.DefaultOntimizeDaoHelper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Lazy;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
-import java.util.HashMap;
+
 import java.util.List;
 import java.util.Map;
+
 
 @Lazy
 @Service("PackService")
@@ -38,6 +32,7 @@ public class PackService implements IPackService {
     private final GuideCitiesService cityService;
     private final ImagePackService imagePackService;
     private final ClientService clientService;
+
 
 
 
@@ -57,7 +52,7 @@ public class PackService implements IPackService {
     @Override
     public EntityResult packQuery(Map<String, Object> keyMap, List<String> attrList)
             throws OntimizeJEERuntimeException {
-        return this.daoHelper.query(this.packDao, keyMap, attrList);
+        return this.daoHelper.query(this.packDao, keyMap, attrList, this.packDao.PCK_MULTI_QUERY);
     }
 
     /**
@@ -69,8 +64,13 @@ public class PackService implements IPackService {
      */
     @Override
     public EntityResult packClientQuery(Map<String, Object> keysValues, List<String> attributes) throws OntimizeJEERuntimeException {
-        keysValues.put(ClientDao.CLIENT_ID, getClientId());
-        return this.daoHelper.query(this.packDao, keysValues, attributes);
+        keysValues.put(ClientDao.CLIENT_ID, clientService.getClientId());
+        return this.daoHelper.query(this.packDao, keysValues, attributes, this.packDao.PCK_MULTI_QUERY);
+    }
+
+    @Override
+    public EntityResult allPacksQuery(Map<String, Object> keysValues, List<String> attributes) throws OntimizeJEERuntimeException {
+        return this.daoHelper.query(this.packDao, keysValues, attributes, this.packDao.PCK_ALL_QUERY);
     }
 
     @Override
@@ -101,28 +101,5 @@ public class PackService implements IPackService {
         return this.daoHelper.delete(this.packDao, keyMap);
     }
 
-    /**
-     * Gets logged client ID
-     * @return Client ID
-     */
-    public int getClientId() {
-        //Gets client object
-        Object client = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
 
-        //Gets logged user id
-        int userId = (int) ((UserInformation) client).getOtherData().get(UserDao.USR_ID);
-
-
-        List<String> qKeys = new ArrayList<String>();
-        qKeys.add(ClientDao.CLIENT_ID);
-
-        Map<String, Object> emptyMap = new HashMap<>();
-        emptyMap.put("CL." + ClientDao.USR_ID, userId);
-
-
-
-        EntityResult clientEr = clientService.clientQuery(emptyMap, qKeys);
-        ArrayList<Integer> al = (ArrayList<Integer>) clientEr.get(ClientDao.CLIENT_ID);
-        return al.get(0);
-    }
 }
